requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.4.0
autopep8==2.3.1
backports.zoneinfo==0.2.1
bcrypt==4.2.0
certifi==2024.7.4
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
passlib==1.7.4
psycopg==3.2.1
psycopg2==2.9.9
pyasn1==0.6.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
ruff==0.6.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.23
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.3
watchfiles==0.22.0
websockets==12.0
zipp==3.20.2
